pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "plot2svg"
version = "0.1.0"
description = "PNG to component-editable SVG MVP pipeline"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "pillow>=10.0",
    "rapidocr-onnxruntime>=1.3",
    "onnxruntime>=1.16",
]

[project.optional-dependencies]
gpu = ["onnxruntime-gpu>=1.16"]
app = ["gradio>=4.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
plot2svg = "plot2svg.cli:main"
plot2svg-app = "plot2svg.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
